% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_row_col_for_grid_item.R
\name{get_row_col_for_grid_item}
\alias{get_row_col_for_grid_item}
\title{Find the row and column for a specific grid item in sample}
\usage{
get_row_col_for_grid_item(grid_item_in, grid_type, grid_file)
}
\arguments{
\item{grid_item_in}{An integer representing the specific item for which the row and column in grid are desired}

\item{grid_type}{An integer representing the type of grid, for grid types already in \code{GMOdetectoR/grids/}}

\item{grid_file}{A path to a grid file if using a grid type not already in \code{GMOdetectoR/grids/}}
}
\value{
A list with values for the row and column of the grid item
}
\description{
Find the row and column for a specific grid item in sample
}
\examples{
\dontrun{get_row_col_for_grid_item(grid_item = grid_item, grid_type = grid_type)}
}
