% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_image}
\alias{load_image}
\title{Load a hyperspectral image}
\usage{
load_image(image_path)
}
\arguments{
\item{image_path}{the path to a hyperspectral image to be loaded as a hyperSpec object}
}
\value{
A hyperSpec object with wavelengths added as column names for the spectra
}
\description{
Load a hyperspectral image
}
\examples{
\dontrun{load_image(image_path)}
}
