% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_image.R
\name{normalize_image}
\alias{normalize_image}
\title{Normalize a reduced image}
\usage{
normalize_image(image_spectrum_table, chroma_table)
}
\arguments{
\item{image_spectrum_table}{an image spectrum table processed by \code{reduce_image}}

\item{chroma_table}{an image spectrum table for a standard, also processed by \code{reduce_image}}
}
\value{
an image spectrum table which has been normalized according to the standard
}
\description{
Normalize a reduced image
}
\examples{
\dontrun{image_spectrum_table <- normalize_image(image_spectrum_table, chroma_table)}
}
