% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_image.R
\name{reduce_image}
\alias{reduce_image}
\title{Extract specific wavelengths from hyperSpec object to prepare for false color image}
\usage{
reduce_image(
  image_in = image_spectrum,
  FP = "unspecified",
  wavelength_make_green,
  wavelength_make_red
)
}
\arguments{
\item{image_in}{A hyperSpec object obtained from \code{load_image}}

\item{FP}{An optional parameter (set to "unspecified" by default) that allows the user to select specific FPs with known fluorescence peaks rather than inputting the peak wavelength itself}

\item{wavelength_make_green}{If option FP is not submitted, the wavelength to which green false color will be applied}

\item{wavelength_make_red}{If option FP is not submitted, the wavelength to which red false color will be applied}
}
\value{
An image spectrum table containing five columns: rows, cols, r, g, b
}
\description{
Extract specific wavelengths from hyperSpec object to prepare for false color image
}
\examples{
\dontrun{reduce_image(image_in, FP="DsRed")}
}
